@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 class NoGlowingOverscrollIndicatorBehavior extends ScrollBehavior {
   @override
-  Widget buildViewportChrome(
-      BuildContext context, Widget child, AxisDirection axisDirection) {
+  Widget buildOverscrollIndicator(
+      BuildContext context, Widget child, ScrollableDetails details) {
     return child;
   }
 }
